sparse: Handle larger FILL and DONT_CARE chunks
The unsparsed chunk size is expressed in units of blk_sz. For RAW chunks
this is then limited to (just under) 32-bits due to the total_sz
(sparsed chunk size), but FILL and DONT_CARE can be much bigger.

But doing the size calculation on the 32-bit integers overflows and any
upper bits are lost, and the unsparsed result is wrong. So, switch this
to 64-bit instead.

It's unclear what happens if we pass a larger num_sectors to the
programmer, so this limitation is left, and checked for.

Fixes: 02c008adfd95 ("add support sparse attribute")

+++ sparse.h
@@ -60,7 +60,7 @@ int sparse_header_parse(int fd, sparse_header_t *sparse_header);
  * Returns the chunk type on success, or an error code otherwise.
  */
 int sparse_chunk_header_parse(int fd, sparse_header_t *sparse_header,
-			      unsigned int *chunk_size,
+			      uint64_t *chunk_size,
 			      uint32_t *value, off_t *offset);
 
 #endif
